Add tests for ROMOnly, MBC1, and MBC2. Fix bugs found during testing!

+++ gb-emu-lib/MBC1_MBC.cpp
@@ -57,7 +57,7 @@ byte MBC1_MBC::ReadByte(const ushort& address)
             targetBank |= (m_ROMRAMBankUpper << 4);
         }
 
-        byte target = (address - 0x4000);
+        ushort target = (address - 0x4000);
         target += (0x4000 * targetBank);
         return m_ROM[target];
     }
